SalesDataGeneration/requirements.txt
asttokens==3.0.1
azure-core==1.37.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
azure-storage-file-datalake==12.22.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jupyter-client==8.7.0
jupyter-core==5.9.1
matplotlib-inline==0.2.1
msal==1.34.0
msal-extensions==1.3.1
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.5.1
prompt-toolkit==3.0.52
psutil==7.1.3
pure-eval==0.2.3
pycparser==2.23
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
